requirements.txt
Pillow
numpy
pytest
sphinx
tensorflow